Convert image to stips
In this file, using thresholding and openCV, we will cover a lot of steps as denoted in the file. This is the first version and will be updated when additional steps are done.

Goal: Convert an image of a salamander to a matrix, where only the stips of the salamander induce non-zero elements in the matrix.