Smoother middle shape
For now, the program will use an image of a salamander. Try to detect the position of the salamander. This, after being smoothend up, will be drawn on a black background. Thus, we actually have an area were the salamander is going to be.

Next is to make a function that can combine this with the adaptive thresholds. Thus we only need to look at the now good marked position, were the salamander will be. So hopefully not too much background!