Pre- and Post-processing steps
Now there are additional pre- and post-precessing steps added (e.g. extra blurring and thresholding). These are highly important regarding to the literature. Indeed, it gives better results.

Furthermore, a lot of experiments have been done and some parameters have been improved.

Also post-processing gives very clear dots on the body of the salamander (mostly when the starting image was of not bad quality).

Next up is to identify these clear dots and to improve the existing code.

Still issues:
1. some images have very bad quality, even after thresholding.
2. The program assumes that the belly of the salamander is in the middle of the image, this is not always the case.