Detection of the good dots & matrix representation
Step 7 is done. We removed a lot of false positives and we can detect the good stips. Then, furthermore we have transformed the image to a matrix and showed this matrix in a nice way.

Next up: Step 8 which requires a lot of improvements and generalizations for worse quality images & some more testing.