Major update: isolating the salamander
In this major update an extra file "isolate_salamander" is added. This file uses color segmentation and morphological operations and clever tricks to only detect the salamander.
Short: we remove everything with different color than that of a salamander. Then we do some morphological operations to improve the structure of the new image. Then we need to remove background (which had the same color). So we use the fact that te salamander is centered in the middle. That it is big and that the box around the salamander fully surrounds the salamander. This, and a combination of some more observations and carefully putting them together lead to this commit and new version, where we most of the time succesfully detect the salamander and no background. (Even in some cases, if the salamander is connected with background, then we can cut this off at the precise point by morphological operations).

This means that step 4 in convert_image_to_stips can be replaced by this new file. Therefore a boolean is added to set which version you prefer to use.

REMARK: The new version gave near perfect results with the 2019 and 2024 images on all the images! The old version could not do this.

USAGE: You can still run convert_image_to_stips but be aware, give a boolean as input for the version. You can also immediatly run the new file with the function "isolate_salamander" This gives you the image of the isolated salamander as a return.

ASSUMPTIONS for new version: The background of the image is not yellow and brown and the salamander is near the middle of the image. (These are minimal assumptions, knowing that the pictures (from Nick) will be taken with a new setup.

ISSUES: they arrize when the assumptions are not satisfied or if the quality is very poor or if the salamander is very very close to the box it is in. In these cases, we have might have a bad or good result, but probably not optimal result. But this are standard problems that you will face in probably every method.