Compare dot patterns
This push includes a new document to compare dot patterns. It is far from ready, but this is a first draft. I follow the literature (these are mentiont at the start of the Python file).

Again, this file works with steps and I will push once certain steps are finished.

Also there is a mini change in the document convert_image_to_stips, but this is nothing new.