Completion of steps 0, 1, 2 and 3
This commit includes the completion of steps 0 to 3. In these steps we extract the coordinates of the dots of the haar cascade method. Then we make some coordinate changes. Then we compute all possible triangles, filter these triangles and compute some extra information about them. Lastly there is also an option to plot a histogram regarding the R-values.