Step 5 done: matching triangles and reducing false matches
In this update, we have added some tweaks to steps 0-4 for better usage. Then, step 5 is finished, in this step we try to reduce the number of false matches between the triangles. This is done with some calculations which can be found in the code itself.

Next is to identify the matching dots, using the matching triangles.