Step 6 done, matching points
Step 6 is completed, in this step we match points, based on the matching triangles and based on a clever voting system. This voting system essentially filters out a lot of the false matched points.

Next up: we continue with some extra stuff that makes the method more robust and reliable.