Extra cropping to only detect belly
For the images of 2024, as of now, only the belly (and somethimes part of the head but that will change in the future) will be detected. For other years it is a bit less good but most of the time worth using over the old version. If you wish to still get the old isolate salamander image, this is possible, when reading the documentation carefully.