Cutting of the head works
With this commit, isolate salamander works a lot better, we (mostly) cut of the tail, paws and head of the salamander and only leave the belly in place. We also rotate the body of the salamander such that it is either perfectely horizontal or perfectely vertical. This helps with cutting of the limbs and the head.