New isolate_salamander + facade
There is a fresh, new version of isolate_salamander! Now it also uses pose estimation to aim for better results! (The old version can still be used and will be used when the new version does not work properly).

Also there is a big change in the facade. The complex pipeline structure of cropping the salamander, using different tools is implemented, though the looks are very messy.

For safety reasons to fall back on when needed, I still left the old isolate salamander in here, but in the future it will be deleted.